VAS updated

+++ server/graphql/types/worksheet/loaded-worksheet-detail.ts
@@ -3,6 +3,6 @@ import { gql } from 'apollo-server-koa'
 export const LoadedWorksheetDetail = gql`
   input LoadedWorksheetDetail {
     name: String!
-    loadedQty: Int!
+    loadedQty: Float!
   }
 `
